requirements.txt
flask
flatlib
swisseph
gTTS
pyttsx3
flask-cors
rembg
pillow
